pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "librouteros>=3.4.1",
    "paramiko>=4.0.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
